requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
pillow==10.3.0
pytesseract==0.3.13
gunicorn==22.0.0
